moved OGIP to untils
