+++ threeML/plugins/OGIPLike.py
@@ -1,13 +1,12 @@
 import pandas as pd
+from astromodels.utils.valid_variable import is_valid_variable_name
 
 from threeML.plugins.DispersionSpectrumLike import DispersionSpectrumLike
 from threeML.plugins.SpectrumLike import SpectrumLike
 from threeML.plugins.XYLike import XYLike
 from threeML.plugins.spectrum.pha_spectrum import PHASpectrum
-from threeML.plugins.OGIP.pha import PHAII
-from astromodels.utils.valid_variable import is_valid_variable_name
-
-from threeML.plugins.OGIP.pha import PHAWrite
+from threeML.utils.OGIP import PHAII
+from threeML.utils.OGIP import PHAWrite
 
 __instrument_name = "All OGIP-compliant instruments"
