+++ threeML/test/test_spectrum_class.py
@@ -1,13 +1,12 @@
+import numpy as np
+import os
 import pytest
-from threeML.plugins.spectrum.binned_spectrum import BinnedSpectrum, BinnedSpectrumWithDispersion, ChannelSet
-from threeML.plugins.SpectrumLike import SpectrumLike
-from threeML.plugins.DispersionSpectrumLike import DispersionSpectrumLike
-from threeML.plugins.OGIP.response import OGIPResponse
 from astromodels import Powerlaw, PointSource, Model
 
-import numpy as np
-
-import os
+from threeML.plugins.DispersionSpectrumLike import DispersionSpectrumLike
+from threeML.plugins.SpectrumLike import SpectrumLike
+from threeML.plugins.spectrum.binned_spectrum import BinnedSpectrum, BinnedSpectrumWithDispersion, ChannelSet
+from threeML.utils.OGIP import OGIPResponse
 
 __this_dir__ = os.path.join(os.path.abspath(os.path.dirname(__file__)))
