+++ threeML/utils/data_builders/time_series_builder.py
@@ -1,29 +1,25 @@
-import numpy as np
-
-from threeML.utils.time_series.time_series import TimeSeries
-from threeML.io.file_utils import file_existing_and_readable
-from threeML.exceptions.custom_exceptions import custom_warnings
-
-from threeML.plugins.OGIPLike import OGIPLike
-from threeML.plugins.OGIP.pha import PHAWrite
-from threeML.plugins.spectrum.binned_spectrum import BinnedSpectrum, BinnedSpectrumWithDispersion
-
-from threeML.utils.data_builders.fermi.gbm_data import GBMTTEFile, GBMCdata
-from threeML.utils.data_builders.fermi.lat_data import LLEFile
-
-from threeML.utils.time_series.event_list import EventListWithDeadTime, EventListWithLiveTime, EventList
-from threeML.utils.time_series.binned_spectrum_series import BinnedSpectrumSeries
-from threeML.plugins.OGIP.response import InstrumentResponse, InstrumentResponseSet, OGIPResponse
-from threeML.plugins.SpectrumLike import SpectrumLike, NegativeBackground
-from threeML.plugins.DispersionSpectrumLike import DispersionSpectrumLike
-from threeML.utils.time_interval import TimeIntervalSet
-from threeML.utils.stats_tools import Significance
-from threeML.io.progress_bar import progress_bar
-
 import copy
-import re
 
 import astropy.io.fits as fits
+import numpy as np
+import re
+
+from threeML.exceptions.custom_exceptions import custom_warnings
+from threeML.io.file_utils import file_existing_and_readable
+from threeML.io.progress_bar import progress_bar
+from threeML.plugins.DispersionSpectrumLike import DispersionSpectrumLike
+from threeML.plugins.OGIPLike import OGIPLike
+from threeML.plugins.SpectrumLike import SpectrumLike, NegativeBackground
+from threeML.plugins.spectrum.binned_spectrum import BinnedSpectrum, BinnedSpectrumWithDispersion
+from threeML.utils.OGIP import InstrumentResponse, InstrumentResponseSet, OGIPResponse
+from threeML.utils.OGIP import PHAWrite
+from threeML.utils.data_builders.fermi.gbm_data import GBMTTEFile, GBMCdata
+from threeML.utils.data_builders.fermi.lat_data import LLEFile
+from threeML.utils.stats_tools import Significance
+from threeML.utils.time_interval import TimeIntervalSet
+from threeML.utils.time_series.binned_spectrum_series import BinnedSpectrumSeries
+from threeML.utils.time_series.event_list import EventListWithDeadTime, EventListWithLiveTime, EventList
+from threeML.utils.time_series.time_series import TimeSeries
 
 
 class BinningMethodError(RuntimeError):
